Fixed shift warning with 32-bit builds.

+++ src/core/WorkItemBuiltins.cpp
@@ -1974,7 +1974,7 @@ namespace oclgrind
           {
             uint64_t a = UARGV(0, i);
             uint64_t b = UARGV(1, i);
-            uint64_t c = (a > UINT64_MAX-b) ? (1L<<63) : 0;
+            uint64_t c = (a > UINT64_MAX-b) ? (((uint64_t)1)<<63) : 0;
             result.setUInt(((a + b) >> 1) | c, i);
             break;
           }
@@ -2214,7 +2214,7 @@ namespace oclgrind
           {
             uint64_t a = UARGV(0, i);
             uint64_t b = UARGV(1, i);
-            uint64_t c = (a > UINT64_MAX-(b+1)) ? (1L<<63) : 0;
+            uint64_t c = (a > UINT64_MAX-(b+1)) ? (((uint64_t)1)<<63) : 0;
             result.setUInt(((a + b + 1) >> 1) | c, i);
             break;
           }
